feat: getGrantingTicket
