+++ server/src/main/kotlin/fi/oph/kitu/oppijanumero/HttpClientConfig.kt
@@ -1,0 +1,16 @@
+package fi.oph.kitu.oppijanumero
+
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import java.net.http.HttpClient
+import java.time.Duration
+
+@Configuration
+class HttpClientConfig {
+    @Bean
+    fun HttpClient(): HttpClient =
+        HttpClient
+            .newBuilder()
+            .connectTimeout(Duration.ofSeconds(10))
+            .build()
+}
